Finished Problems 03 and Problems04 in just 1 night

+++ CSC-190 Object-Oriented Soft Dev/Self Study/Lesson 9 Collections Framework/Problems01 Easy (Pure Logic)/Main.java
@@ -1,5 +1,5 @@
 // Creation Date: June 28, 2026. at 4:02 AM
-// Last Modified: June 29, 2026. at 11:17 AM
+// Last Modified: June 30, 2026. at  1:42 AM
 
 import java.util.ArrayList;
 
@@ -20,7 +20,7 @@ public class Main {
         int result = 0;
 
         for (Integer i:arr) {
-            if ((double)i%2 == 0) { // If dividing 2 leaves 0 remainder, then do this (This means that if a number is divided by 2 returns a whole number, then it's even) NOTE: in order to get that decimal point, we must convert the data type into float, long, or double in order to store decimal numbers
+            if (i%2 == 0) { // If dividing 2 leaves 0 remainder, then do this (This means that if a number is divided by 2 returns a whole number, then it's even) NOTE: in order to get that decimal point, we must convert the data type into float, long, or double in order to store decimal numbers
                 result += i;
             }
         }
@@ -28,3 +28,13 @@ public class Main {
         return result;
     }
 }
+
+
+// Problem 1 — Easy (Pure Logic)
+// Given an ArrayList<Integer>, return the sum of all even numbers in the list.
+//
+// Input:  [1, 2, 3, 4, 5, 6]
+// Output: 12  (2 + 4 + 6)
+//
+// Input:  [1, 3, 5, 7]
+// Output: 0  (no even numbers)
